added comms and confusion

+++ src/common/Supplier.java
@@ -1,16 +1,16 @@
 package common;
 
 public class Supplier extends Model {
-    private Integer distance;
+    private double distance;
     //items sold??
 
-    public Supplier(String name, int distance) {
+    public Supplier(String name, double distance) {
         this.setName(name);
         this.distance = distance;
     }
 
     //Sets & Gets
-    public int getDistance() {
+    public double getDistance() {
         return distance;
     }
